GDA-5883 Scan and ScanDataPoint now use an int for scan number

+++ uk.ac.gda.core/src/gda/scan/PassthroughScanAdapter.java
@@ -168,7 +168,7 @@ public class PassthroughScanAdapter implements NestableScan {
 	}
 
 	@Override
-	public Long getScanNumber() {
+	public int getScanNumber() {
 		return delegate.getScanNumber();
 	}
